+++ iOSswiftUIa11yTechniques/iOSswiftUIa11yTechniques/Prototypes/PrototypesView.swift
@@ -47,9 +47,6 @@ struct PrototypesView: View {
             NavigationLink(destination: MarkdownView()) {
                 Text("Markdown View Test")
             }
-            NavigationLink(destination: WebView()) {
-                Text("Web View Apple System CSS Font Test")
-            }
         }
         .navigationTitle("Prototypes")
     }